More fixes for production
The fonts were failing to load. Viewing the bundled/minified CSS file
revealsed that minification failed for the @import statements. Here's
the problem: @import statements must be the first things in a CSS file.
So if a CSS file contains @imports at the top, but it's not the first
CSS file in the bundle, then the @imports won't be at the top of the
bundled file. My solution was to make a separate CSS file that just
contains @imports, and put that at the start of the bundle. The rest of
my CSS (in the other file) remained at the end of the bundle.